# build: Add Develocity configuration and enable remote build cache