--- a/gradle.properties
+++ b/gradle.properties
@@ -2,3 +2,5 @@ group=org.eclipse.edc
 version=0.12.0-SNAPSHOT
 edcScmUrl=https://github.com/eclipse-edc/Connector.git
 edcScmConnection=scm:git:git@github.com:eclipse-edc/Connector.git
+
+org.gradle.caching=true