@ribafish 2025-03-14
Note: I modified the PR title to include `build:`, as I saw a [check failed](https://github.com/eclipse-edc/Connector/actions/runs/13807454363)