@ribafish 2025-04-04
Hi @paullatzelsperger, the parent issue https://github.com/eclipse-edc/Connector/issues/4846 was closed by the bot due to it being stale. Is this still being considered, or do you perhaps have any questions that I could answer?