@paullatzelsperger 2025-04-07
> Hi @paullatzelsperger, the parent issue #4846 was closed by the bot due to it being stale. Is this still being considered, or do you perhaps have any questions that I could answer?

@ribafish I was OOF for a bit, let me catch up with the team, but IIRC last we spoke the consensus was that we want to gather experience from other projects (namely Tractus-X) first, and then make a final decision whether we want to use Develocity or not.